changed the longs to ints (type of volume in the volume_changed callback).

+++ src/phoneui-utils-sound.c
@@ -25,7 +25,7 @@ static struct SoundControl controls[SOUND_STATE_INIT][CONTROL_END];
 
 /* The sound cards hardware control */
 static snd_hctl_t *hctl = NULL;
-static void (*_phoneui_utils_sound_volume_changed_callback) (enum SoundControlType type, long volume, void *userdata);
+static void (*_phoneui_utils_sound_volume_changed_callback) (enum SoundControlType type, int volume, void *userdata);
 static void *_phoneui_utils_sound_volume_changed_userdata = NULL;
 static void (*_phoneui_utils_sound_volume_mute_changed_callback) (enum SoundControlType type, int mute, void *userdata);
 static void *_phoneui_utils_sound_volume_mute_changed_userdata = NULL;
@@ -593,7 +593,7 @@ _phoneui_utils_sound_element_cb(snd_hctl_elem_t *elem, unsigned int mask)
 }
 
 int
-phoneui_utils_sound_volume_change_callback_set(void (*cb)(enum SoundControlType, long, void *), void *userdata)
+phoneui_utils_sound_volume_change_callback_set(void (*cb)(enum SoundControlType, int, void *), void *userdata)
 {
 	_phoneui_utils_sound_volume_changed_callback = cb;
 	_phoneui_utils_sound_volume_changed_userdata = userdata;
